requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.78.0
h11==0.13.0
idna==3.3
numpy==1.23.0
Pillow==9.2.0
pydantic==1.9.1
python-dotenv==0.20.0
python-multipart==0.0.5
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
torch==1.12.0
typing_extensions==4.3.0
uvicorn==0.18.2
